requirements.txt
haystack-ai==2.0.0b4
streamlit==1.21.0
markdown
st-annotated-text
python-dotenv
mastodon-fetcher-haystack